creation of viewextension
