+++ src/DynamoCoreWpf/Views/Core/DynamoView.xaml.cs
@@ -95,6 +95,8 @@ namespace Dynamo.Controls
 
         private FileTrustWarning fileTrustWarningPopup = null;
 
+        public ShortcutToolbar ShortcutBar { get { return shortcutBar; } }
+
         /// <summary>
         /// Constructor
         /// </summary>
