+++ src/NotificationCenterViewExtension/NotificationCenter_ViewExtensionDefinition.xml
@@ -1,0 +1,5 @@
+<ViewExtensionDefinition>
+  <AssemblyPath>..\NotificationCenterViewExtension.dll</AssemblyPath>
+  <TypeName>Dynamo.NotificationCenter.ViewExtension</TypeName>
+  <IsEnabled>True</IsEnabled>
+</ViewExtensionDefinition>
