+++ src/NotificationCenterViewExtension/View/NotificationUI.xaml
@@ -1,11 +1,10 @@
-﻿<Popup x:Class="Dynamo.Wpf.Views.Notifications.NotificationUI"
-            xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+﻿<Popup x:Class="Dynamo.NotificationCenter.View.NotificationUI"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
        xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-       xmlns:local="clr-namespace:Dynamo.Wpf.Views.FileTrust"
-       xmlns:ui="clr-namespace:Dynamo.UI"
-       xmlns:controls="clr-namespace:Dynamo.Controls"
+       xmlns:ui="clr-namespace:Dynamo.UI;assembly=DynamoCoreWpf"
+       xmlns:controls="clr-namespace:Dynamo.Controls;assembly=DynamoCoreWpf"
        xmlns:p="clr-namespace:Dynamo.Wpf.Properties;assembly=DynamoCoreWpf" 
        xmlns:fa="http://schemas.fontawesome.io/icons/" 
        xmlns:wv2="clr-namespace:Microsoft.Web.WebView2.Wpf;assembly=Microsoft.Web.WebView2.Wpf"
@@ -74,7 +73,7 @@
         </Path>
 
         <Grid  x:Name="mainPopupGrid" Background="White"                
-               Width="{Binding PopupRectangleWidth}">         
+               Width="{Binding PopupRectangleWidth}">
             <wv2:WebView2  Source="http://localhost:3000/"></wv2:WebView2>
         </Grid>
     </Canvas>
