+++ src/NotificationCenterViewExtension/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Microsoft.Web.WebView2" version="1.0.1245.22" targetFramework="net48" />
+</packages>
